Skip forced mgmt route test if intf is invalid (#13438)
What is the motivation for this PR?
Skip test_forced_mgmt_route_add_and_remove_by_mgmt_port_status in route/test_forced_mgmt_route.py if eth1 does not exist or is being used. This test was failing on Arista platforms when interface eth1 did not exist.

How did you do it?
Added a check to see if eth1 does not exist, and skips the test if so.

How did you verify/test it?
Tested on a t0 setup on a Arista-7260

Signed-off-by: yenlu <yenlu@cisco.com>