Move platform api test to skip list (#13567)
Description of PR
In order to test the api of platform, we need the module sonic_platform, which is not included in vs testbed. Thus, it will get ModuleNotFoundError: No module named 'sonic_platform' on vs. We will skip these tests on vs.

What is the motivation for this PR?
In order to test the api of platform, we need the module sonic_platform, which is not included in vs testbed. Thus, it will get ModuleNotFoundError: No module named 'sonic_platform' on vs. We will skip these tests on vs.

How did you do it?
Move the scripts under folder platform_tests/api from pr_test_scripts.yaml to pr_test_skip_scripts.yaml for caculation.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>