[Golden Config Override] Build test infra to simulate NDM override behavior to test new features (#13477)
Description of PR
Summary: This PR is to build a test insfrasture to simulate the process of Golden Config override with new table or exisitint tables.
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Back port request
 202012
 202205
 202305
 202311
 202405
Approach
What is the motivation for this PR?
There is a need to validate new feature or new tables without change in minigraph parser. Here we setup a test infrastructure to mimic the behavior of Golden Config override process and override ConfigDB with the desired template.

How did you do it?
Define a new way of config reload with override.
Add a common template to generate common feature and add a test case to generate specific table to test.

How did you verify/test it?
E2E test

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>