Added validation of HW revision on mellanox devices in test_show_platform (#13364)
What is the motivation for this PR?
Following the correction of hw-revision in show platform summary command (sonic-net/sonic-buildimage#19098), we add a validation to mellanox devices that the hw-revision is as expected.

How did you do it?

How did you verify/test it?
Compare the output in show platform summary with that specified in the predefined path (VPD_DATA_FILE )

Signed-off-by: yenlu <yenlu@cisco.com>