Update test_plan.py to get token via az cli (#13440)
What is the motivation for this PR?
This is the second part of the work for eliminating client secrets.

This change depends on merge of #13435.

How did you do it?
This change updated the test_plan.py to get access token by az cli. For the az cli command to work, the task calling test_plan.py must be wrapped in an AzureCLI azure pipeline task.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>