[dhcp_server_test] Enable autorestart before test dhcp server container autorestart (#13540)
Description of PR
Summary:
Fixes # (issue)
test_pretest.py disables autorestart for all features, so if we want to test dhcp server autorestart we need to enable it.

Approach
What is the motivation for this PR?
test_pretest.py disables autorestart for all features, so if we want to test dhcp server autorestart we need to enable it.

How did you do it?
Enable autorestart before test

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>