Add onboarding dualtor PR checker (#13591)
What is the motivation for this PR?
In this PR, we add a new PR checker called onboarding dualtor. This new PR checker is for onboarding dualtor speific test scripts. Also, we modify the pytest mark from t0 to dualtor of two dualtor specific scripts to make it more clear.

How did you do it?
Add a new PR checker called onboarding dualtor. Also, we modify the pytest mark from t0 to dualtor of two dualtor specific scripts to make it more clear.

Signed-off-by: yenlu <yenlu@cisco.com>