Fix test_neighbor_mac on t1-lag and t1-64-lag (#13314)
* Fix test_neighbor_mac on t1-lag and t1-64-lag

A recent change sets this test to t1-only:
https://github.com/sonic-net/sonic-mgmt/pull/12952

The test can pass on topo t1 without any change.

For t1-lag or t1-64-lag, the Ethernet0 to test is in portchannel, so
IP address cannot be set to it. The fix here is to delete the intf
from the portchanenl before testing it.

Signed-off-by: yenlu <yenlu@cisco.com>