Add functionality to save techsupport on vSONiC neighbors, and use it for retry count (#13605)
What is the motivation for this PR?
For test cases where there is some configuration done on the neighbor devices, or some state is checked on the neighbor devices, if there is a test case failure, it would be useful to get the logs on that device for debugging purposes. However, there's no fixture currently that does this (there is a fixture that automatically collects techsupport on the DUT itself).

How did you do it?
Add a fixture to collect the techsupport from neighbor VMs if they are vSONiC neighbors if the test case fails. Also, modify two test cases in tests/pc/test_retry_count.py to have it collect information from the neighbors if it fails. This information will be extremely helpful in debugging failures.

How did you verify/test it?
Modified the test function to always collect techsupport, then run it on a KVM t0-116 topo with vSONIC neighbors, and verify that techsupport got collected.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>