Fix arp/test_stress_arp.py to not try and populate 100% of available ipv6 resources
Previously the test attempts to use up 100% of available ipv6 resources. The test is flaky and fails occasionally.

A previously fixed bug (967459) also exists for the ipv4 version of this test, and reducing the usage of resources by 1/3 fixed it as it is highly likely a hash collision problem.

Here we attempt to use the same solution even though historically the fail pattern and % of the ipv6 version of the test does not really match the ipv4 version.