Add debug message for debugging script `test_stress_routes.py`. (#13492)
What is the motivation for this PR?
We got flaky issue ConnectionError(ProtocolError('Connection aborted.', ConnectionResetError(104, 'Connection reset by peer'))) in script test_stress_routes.py in baseline test. We can not reproduce this issue in local environment. The logs of announcing routes are stored on DUT, in order to save these logs on test server and downloaded by elastictest, we use function config_module_logging to achieve this.

How did you do it?
Use function config_module_logging to get the logs of announcing route ftom DUT to server, and store them in the folder logs.