Add new generic hash test into PR checker.  (#13679)
What is the motivation for this PR?
In PR #12569, we got a new test script hash/test_generic_hash.py, which is not included in our onboarding PR checker. In this PR, we add it into onboarding T0 and T1 PR checker to verify if this test can run in PR test successfully.

How did you do it?
Add test script hash/test_generic_hash.py into onboarding T0 and T1 PR checker.