Add module `test_gnmi_countersdb.py` into onboarding PR checker (#13833)
What is the motivation for this PR?
In PR #13696, there is a new added module test_gnmi_countersdb.py. In this PR, add this new module into onboarding PR checker. If it is stable enough, we will move it to the standard PR checker.

How did you do it?
Add the new added module test_gnmi_countersdb.py into onboarding T0 and T1 PR checker.

Signed-off-by: yenlu <yenlu@cisco.com>