Add a condition while skipping the script `test_pktgen.py` (#13807)
What is the motivation for this PR?
Script test_pktgen.py has a known issue, and will skip on all testbeds using conditional marks. Previously, there is only the reason in condition marks. In this PR, we add an issue as the condition, and when the issue is closed, this script will not be skipped. Also, we add this script into PR checkers.

How did you do it?
Add an issue as the condition in conditional marks.
Add script test_pktgen.py into PR checkers.

Signed-off-by: yenlu <yenlu@cisco.com>