[QoS]qos_yaml j2C+ changes for new _vsq thresholds (#13069)
What is the motivation for this PR?
The new   MMU settings to enhance performance for RDMA traffic in production.
Hence the qos_params needs to be tweaked according to the set buffer profiles.

However, the existing sonic-mgmt LossyQueueTest  doesn't fairly verify the buffer threshold for headroom for Lossy traffic. As per the new vsq profile setting the XOFF FADT threshold/PG is way lesser than the Nominal headroom, which limits it to not utilize the headroom buffer completely and send pause frames before reaching the MAX headroom limit.
Either the test case needs to be improvised by adding more source ports or a new test case should be added to verify the Lossy queue traffic at PG level

Signed-off-by: yenlu <yenlu@cisco.com>