Wait for create acl table/rule msg in syslog using scripts (#13858)
In case of T2 profile, it takes longer time in some cases for create acl table message to appear in syslog. Adding a wait_until proc to check for the message. @abdosi has context of the issue and the issue has been discussed in chassis community meeting.

What is the motivation for this PR?
In case of T2 profile, it takes longer time in some cases for create acl table message to appear in syslog. Adding a wait_until proc to check for the message

How did you do it?
How did you verify/test it?
Verified it on T2 profile with Cisco chassis