Deprecate ceosimage_saskey for downloading cEOS image (#13840)
What is the motivation for this PR?
To download cEOS image, the URL has to be composed using multiple components including ceosimage_saskey. This is complicated and unnecessary.

How did you do it?
This change simplified the configuration of URL for downloading cEOS image. Just configure variable ceos_image_url in ansible/group_vars/all/ceos.yml is enough.

This change also enhanced the code logic to support configuring ceos_image_url as a string or a list of strings. When it is a string, the value is considered as a single URL. When it is a list of strings, each item in the list is considered as a URL. Playbook will try the images URLs one by one until downloading is successful.

The related document is updated too.

How did you verify/test it?
Tested add-topo.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>