Add script test_show_subinterface.py into PR checker (#13876)
What is the motivation for this PR?
In test script test_show_subinterface.py, we want to create the subinterfaces and check the status of these new created ports. While testing, we find that the new created ports will not inherit the speed of their ancestor, which is not expected. We suspect it is an image issue, so, we create an issue and use conditional mark to skip this case on vs testbed temporarily. As this case is skipped for now, we add it into PR checker, and after the issue is fixed, the case can run in PR test successfully.

How did you do it?
Create an issue, and use conditional mark to skip this case on vs testbed temporarily.

Signed-off-by: yenlu <yenlu@cisco.com>