sonic-mgmt: reduce fdb test time for large topos
Standalone topos with many interfaces, such as
64 or more, may result in individual test runs
from test_dfb.py that are longer in duration
than comparatively smaller topos by orders of
magnitude, making the tests in test_fdb.py
impractical to run in such cases.

This patch reduces the mac count for larger 
standalone topos from 10 to 2, similar to what
has been done in the past for other topos that
do not scale well for this test.