Add a field `IsSkipped` to indicate if a script is skipped in PR checker.  (#13894)
What is the motivation for this PR?
We already have the script to validate and trace the test scripts in PR checker. In this PR, we enhance this script, add a field to indicate if this script is recorded in our skip list.

How did you do it?
Add a filed to indicate if the script is recorded in our skip list.