update detect time in test_pfcwd_timer_accuracy (#13902)
What is the motivation for this PR?
The test case failed in nightly build.
Should be related to the PR #12733, which decrease the detect sleep time for test_pfcwd_timer_accuracy

How did you do it?
Enlarge the detect sleep time

How did you verify/test it?
Run original case

Signed-off-by: yenlu <yenlu@cisco.com>