Add Run test2 step if Run test step timeout
This change is to address the "az login" timeout after 24 hours issue.

The previous fix to run az login again using System.AccessToken is not working. The reason is that the System.AccessToken has expired when "az login" is executed in test_plan.py.

This change takes advantage of the "timeout" argument of test_plan.py. If test_plan.py has been running for like 20 hours, exit early with a different return code.

In the pipeline yaml, set a special variable POLL_AGAIN to "YES" if test_plan.py returned the timeout RC.

This change added "Run test2" after "Run test" step.

In "Run test2", it firstly checks if POLL_AGAIN is set to "YES". If not, just `exit 0` to skip "Run test2". Otherwise, try to run test_plan.py to poll running state again. This time, the timeout is hardcoded to 16 hours so far.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>