Include hostname and tracking ID when uploading warm-reboot results to Kusto (#11178)
* Include the hostname when generating the post-reboot analysis

After doing a warm/fast reboot, in the post-reboot analysis, include the
hostname of the device. This information will then get uploaded into
Kusto (if enabled). The reason for this is that in case there are some
device-specific oddities that cause some things to take longer on one
device compared to another device, then this will help reveal them. It
also makes it easier to see what device a certain run was done on, in
case there need to be some later debugging effort.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Use the supplied tracking ID instead of a random GUID for reports

For the upgrade path test reports, if a tracking ID was passed in, use
that tracking ID instead of a random GUID. This makes it easier to
correlate this upgrade test report with other data (such as a specific
Azure pipeline run).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Include the tracking ID in the upload

Instead of replacing the report ID (which can still serve as a unique
ID, add a separate column for the tracking ID, which can then include
whatever external ID should be used here. This allows the current
ReportId column to not end up overloaded with two separate meanings.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>