Ignore loganalyzer errors for rsyslog RELP (#13853)
* Ignore errors about rsyslogd w/ librelp not being able to send syslogs

If rsyslogd on the host goes down, and rsyslogd on the containers is
configured to use librelp to forward messages to the host rsyslogd
(instead of UDP), then there will be error messages from the container
rsyslogd about not being able to forward messages.

Ignore these error messages as they are expected when running tests
which may restart rsyslogd.

This is in preparation for sonic-net/sonic-buildimage#18113

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>