Snappi-Issue-13752 Support to enable PFCWD in Forward Mode (#13847)
What is the motivation for this PR?
Providing support to enable PFCWD in forward mode.

How did you do it?
Added a function 'start_pfcwd_fwd' with duthost and asic_value as parameters. If asic_value is not passed, then PFCWD in forward mode will be enabled in global mode.

Function first stops the PFCWD (in whatever mode it is operating), before enabling it in FORWARD mode.

How did you verify/test it?
Tested with function call with and without asic_value and checked if PFCWD was enabled on the DUT. If asic_value is passed, PFCWD is enabled for the specific ASIC in forward mode.