[7215] Set timeout = 180s for config reload cmd (#13922)
What is the motivation for this PR?
For Nokia-7215, 120 seconds is not enough for config reload command. Extend the time to 180s to make testcase stable.

How did you do it?
Set config reload timeout to 180s for Nokia-7215.

How did you verify/test it?
Verified on Nokia-7215 Mx testbed.

Signed-off-by: yenlu <yenlu@cisco.com>