Skip cancel test plan for "NIGHTLY" test plan type (#13923)
What is the motivation for this PR?
This change is to address the "az login" timeout after 24 hours issue.

How did you do it?
The previous fix to run az login again using System.AccessToken is not working. The reason is that the System.AccessToken has expired when "az login" is executed in test_plan.py.

This change takes advantage of the "timeout" argument of test_plan.py. If test_plan.py has been running for like 20 hours, exit early with a different return code.

In the pipeline yaml, check return code of test_plan.py. If it is 0 or 2 (polling timeout), consider it as success.

The cancel test plan step still run with condition "always()". This step is updated to check TEST_PLAN_TYPE, if it is nightly, exit early to skip cancel test plan because the test plan could still be running.

How did you verify/test it?.
Use draft PR to verify the change.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>