Log full url in the getbuild.py tool (#13841)
* Log full url in the getbuild.py tool

When the getbuild.py tool fail with downloading artifacts, it is not easy to quickly root cause.

This change updated the code to log full URL for downloading artifacts. Then when this tool failed, we can quickly do a manual check to find out if the issue is with the URL, server, or permission.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>