[dhcp_server_test] wait lease to be wrote in state_db before check and ignore log error (#13914)
We should wait lease to wrote in the state_db before check and we should ignore error in pattern.*Failed to get port by bridge port.*

What is the motivation for this PR?
The test case failed because lease was not updated to state_db, we need to wait on it.

How did you do it?
Wait on lease to be wrote to state db

How did you verify/test it?
Run test on DUT

Signed-off-by: yenlu <yenlu@cisco.com>