Fix keyerror (#13729)
Description of PR
In sciprt hash/test_generic_hash.py, it needs to get asic type from the dict dut_info. But on kvm testbed, this dict doesn't have the key asic_type, which will cause the KeyError. In this PR, we use the function get to get the value in a dict, and if the key doesn't exist, we will return an empty string.

Approach
What is the motivation for this PR?
In sciprt hash/test_generic_hash.py, it needs to get asic type from the dict dut_info. But on kvm testbed, this dict doesn't have the key asic_type, which will cause the KeyError. In this PR, we use the function get to get the value in a dict, and if the key doesn't exist, we will return an empty string.

How did you do it?
Use the function get to get the value in a dict, and if the key doesn't exist, we will return an empty string.

co-authorized by: jianquanye@microsoft.com