[dualtor-aa] Fix flakiness of arp/test_unknown_mac.py (#13794)
What is the motivation for this PR?
Currently test is picking an ip address that isn't being used for server/ptf interfaces, and assigning this ip address to the ptf interface before pinigng the dut vlan. This will cause dut ARP and MAC tables to be populated for the relevant ip.

In case of dualtor active-active topologies we have concept of nic-simulator, where we'll have an ip address for server nic as well (soc_ipv4 field in 'show mux config') along with ptf interface ip (server_ipv4). And there will be gRPC traffic flowing b/w sonic dut and nic (with soc_ipv4).

So updating the test to skip assigning an ip address for a ptf interface (used to ping vlan ip of dut), that's already assigned to nic interfaces (soc_ipv4), as gRPC traffic can override the ARP table entry that is learned during ping from ptf to sonic dut.

And test is failing sometimes with Failed: Drop counters failed to increment, as the traffic is getting permitted because of MAC table entry (populated due to gRPC traffic).

How did you do it?
How did you verify/test it?
Ran the test with fix on Arista-7260CX3-D108C8 and test is passing.

Signed-off-by: yenlu <yenlu@cisco.com>