Tolerate hash collision when adding entries in `test_stress_arp.py` (#13861)
What is the motivation for this PR?
In script test_stress_arp.py, we will add as much ipv4 arp entries and ipv6 nd entries as we can. We will get this number from crm. But in fact, there is a probability of hash collision, which means we can not add such number entries as we expected. So in this PR, we tolerate this possible and set the percentage of hash collision 1%

Signed-off-by: yenlu <yenlu@cisco.com>