Fixed copp/test_policer testcase for marvell platforms (#13989)
What is the motivation for this PR?
Recently, a change was made in the test_policer testcase in copp_tests.py to retain the default value of cir/cbs for queue4_group3 (default value for mgmt is 300) instead of changing it to 600 on the DUT during the testcase run. The PR for this change is here - #12836. For marvell based platforms like Nokia-7215-T1 and Nokia-7215-A1, the cir/cbs values are supported in steps of 125. Therefore, a value of 300 in the copp_cfg.json will set it to 250 internally since 300 is not in steps of 125. This was confirmed by Marvell - "Per talk, the policer/ traffic metering can work in packet per-second mode. The granularity is in step of 125 pps. So it’s 125, 250 …etc Minimum is 125."
The PPS_LIMIT_MIN and PPS_LIMIT_MAX is therefore set in the testcase to 270,390 respectively which is completely out of range and causes this testcase to fail. Therefore, this PR sets cir/cbs to 250 for Marvell based platforms.

How did you do it?
Added a platform specific check in copp_tests.py to change the PPS_LIMIT to 250 for Nokia platforms.

How did you verify/test it?
Ran test_policer testcase for DHCP, DHCP6, LLDP and UDLD. Verified that the testcase passes.