Fix arp/test_unknown_mac by disabling arp_update (#13410)
What is the motivation for this PR?
Test was flaky. Monitoring tcpdump on the PTF container and the DUT's fdb table shows IPv6 echo is causing mac addresses to be learned on the DUT even though it was previously flushed.

How did you do it?
Disable IPv6.

How did you verify/test it?
Disabling IPv6 will allow the test to consistently past. If IPv6 is re-enabled during the test and packets are sent, it immediately fails.

Any platform specific information?

Signed-off-by: yenlu <yenlu@cisco.com>