Add script `platform_tests/test_reload_config.py` into T0 PR checker.  (#13575)
What is the motivation for this PR?
There are some modifications in this PR:

On kvm testbed, the service watchdog-control can not be loaded because of the reason No module named 'sonic_platform, which will also cause the failed load of service system-health. So I raised an issue about this question and use conditional mark to skip this script temporarily.
After reboot, although check system config passes, it doesn't mean all containers are up. So we add another checker to check if all containers are up after rebooting. After all the containers are up and the system is stable, we will do the next operations.

How did you do it?
Raise an issue and use conditional mark to skip this script temporarily
Add a checker to check if all containers are up.