[dhcp_server_test] fix test failure due to lease was not updated to state db (#14120)
When running nightly test, sometimes, there are test cases failed on verify lease in state db.

After investigation, I found that it's because when ip assigned was expired it will be appended to the last line in lease file
and dhcpservd will choice that line. When two tests randomly select the same port, and the previous test didn't release the ip, the next test may fail, because dhcpservd will only pick the last record of that mac in lease file.

To fix that, we need to release ip assigned after test.

What is the motivation for this PR?
When running nightly test, sometimes, there are test cases failed on verify lease in state db.

How did you do it?
Release ip assigned after test.

How did you verify/test it?
Run tests on testbed and monitor lease file to make sure every lease was released after the test.

Signed-off-by: yenlu <yenlu@cisco.com>