Merged PR 6909: Update 7050cx3 pipelines for 202311 image qualification
Update 7050cx3 pipelines for 202311 image qualification

Have vms28-t0-7050-12 do in-branch 202311 upgrades (with vSONiC), and
vms28-t0-7050-14 do 202305 to 202311 upgrades (with cEOS) along with SAD
test cases. Also, update the iteration count from 18 to 22 for non-SAD
pipelines.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Pipeline update for image qualification.

#### PR Summary
This pull request updates the pipelines for the 7050cx3 series to qualify the 202311 image.
- `.azure-pipelines/sonic-metadata/arista/vms28-t0-7050-14-2.yml`: Added 202311 images and updated default image URLs.
- `.azure-pipelines/sonic-metadata/arista/vms28-t0-7050-14.yml`: Added 202311 images and updated default image URLs.
- `.azure-pipelines/sonic-metadata/arista/vms28-t0-7050-12.yml`: Added 202311 images and updated default image URLs.
- `.azure-pipelines/sonic-metadata/arista/vms28-t0-7050-13.yml`: Updated VSONIC_NEIGHBOR_IMAGE_URL to use 202311 image.