split test_device_checker into to two cases (#13278)
Fix issue:https://redmine.mellanox.com/issues/3923746
When test mocks fan target speed and actual speed, it will cause fan tacho err, because the actual speed is not equal to target speed. This tacho err will exists until the whole test is finished, because we don't recover the mocked fan soft links until the test is finished. Therefore, when mocking the fan dir err, there will be two tc errors, so the  the pwm will be adjusted to 100, and the speed of fan also should be adjusted to full speed accordingly. But usually the fan speed cannot be adjusted to 100 in a short time, before it is adjusted to the full speed, there will be some waring msg such as "WARNING pmon#thermalctld: Fan low speed warning: fan*".

Change-Id: I51239fe41a8b7d7d6bbf7b0877a78116ba11f908