Fix arp/test_stress_arp.py ipv4 and ipv6 inconsistency and bad checks
There are 2 issues with arp/test_stress_arp.py:

1. It is doing a redundant subtraction.
In both variants, it calculates space available as (available - used)
instead of (total size - used), giving an incorrect value for available
space.

2. Not enough allowance/wiggle room for hash collision.
In both variants, it expects only 100 hash collisions which is very
little. For specific SKUs, there may be a further performance impact as
some hash functions is not as performant for incremental values, which
is what is used for this test.

This patch fixes these 2 issues by:
   1. Removing the redundant subtraction.
   2. Change hash collision allowance from 100 to 250.