Recover testbeds while the password is not correct (#14212)
Approach
What is the motivation for this PR?
There are many unhealthy testbeds which password is not incorrcet. And our auto recovery didn't cover this sistuation before. In this PR, we optimise related code. As long as the testbed is unreachable through ssh, not matter which sistuation, we will reimage the testbed from boot loader.

co-authorized by: jianquanye@microsoft.com