[Snappi]: Adding support for breakout links (#14026)
Approach
What is the motivation for this PR?
This PR supports the addition of breakout ports

How did you do it?
How did you verify/test it?
Instead of mentioning Card info , we can dirently mention the fanout port info in the ansible/links file example Port1.1, Port1.2 and so on

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>