fix: add extra wait time for ACL test (#14242)
Description of PR
In the last chassis sync meeting, we agreed on the following changes to acl/test_acl.py to make it more stable on T2 chassis:

Increase the reboot wait time
Increase the wait time for port toggle to 5 min for both up and down
Increase the aclshow -a wait time
In this PR, we have increased the reboot wait time by enabling safe_reboot and increased the wait time for port toggle to 5 min if it is a chassis. The aclshow -a wait time increase has been included in #13858.

Summary:
Fixes # (issue) Microsoft ADO 29182672

Approach
What is the motivation for this PR?
Make acl/test_acl.py more stable on Cisco T2 chassis

How did you do it?
How did you verify/test it?
I ran the updated test on Cisco T2 chassis and made sure the changes are in-place.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>