Wait for create acl table/rule msg in syslog using scripts [202405 branch] (#14255)
Creating this PR for 202405 branch. Same PR (#13858) was merged for master but had conflicts for 202405 branch. Hence opening this one for 202405 branch

In case of T2 profile, it takes longer time in some cases for create acl table message to appear in syslog. Adding a wait_until proc to check for the message. @abdosi has context of the issue and the issue has been discussed in chassis community meeting

What is the motivation for this PR?
In case of T2 profile, it takes longer time in some cases for create acl table message to appear in syslog. Adding a wait_until proc to check for the message

How did you do it?
Verified it on T2 profile with Cisco chassis

How did you verify/test it?
Verified it on T2 profile with Cisco chassis