chore: temporarily disable orchagent check for BFD test (#14261)
Description of PR
As discussed in the last chassis meeting, we want to temporarily disable orchagent CPU check for BFD tests as it is not stable at the moment.

Summary:
Fixes # (issue) Microsoft ADO 29191540

Approach
What is the motivation for this PR?
How did you do it?
How did you verify/test it?
I ran the updated BFD tests and confirmed that orchagent check is disabled

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>