feat: add orchagent usage check in sanity check (#14263)
Description of PR
As discussed in the last chassis meeting, we want to add an orchagent usage check in sanity check to confirm if the orchagent is behaving well in T2 Nightly runs. All other Nightly runs can also benefit from this change.

Summary:
Fixes # (issue) Microsoft ADO 29191540

Approach
What is the motivation for this PR?
We saw the orchagent usage is being very high in the middle of the chassis test, so we wanted to add orchagent usage check to the sanity check so we can see how it behaves in each test.

How did you do it?
How did you verify/test it?
I ran some tests with the updated sanity check to confirm it's working.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>