Fix flaky test_reload_configuration_checks when all processes are not up during swss stop job (#14218)
What is the motivation for this PR?
The testcase test_reload_configuration_checks fails sometimes when swss is stopped after a config reload and some of the critical processes are still coming up. The stop job of swss in the queue is cancelled due to other critical processes still coming up and trying to bring up swss. Hence, we get the error - "Job for swss.service cancelled"

How did you do it?
This PR enhanced the testcase to wait until all the critical processes are up after a config reload and then execute a stop job for swss.

How did you verify/test it?
Ran the testcase 15-20 times to see if it fails.

Signed-off-by: yenlu <yenlu@cisco.com>