[test_bgp_stress_link_flap] case hung sometimes due to memory exhaust (#14163)
What is the motivation for this PR?
The case is flaky
Sometimes the case runs a long time and no response. Especially on kvm device.
Based on current log, it should be related to the memory resource limitation.
The case would create so many threads to flap the neighbor, it would cause kvm device memory exhaust, and same for low memory physical device.
Based on available logs, no obvious memory leak issue.

How did you do it?
The case is for stress link flap, it creates thread per interface to flap.
1: enlarge the delay time for kvm
2: only test one interface for kvm
3: Use one thread to flap all the interfaces for fanout.
4: correct neighbor host
6: add event stop and timeout for thread function to ensure thread exit

How did you verify/test it?
run the case locally and verified using elastictest
https://elastictest.org/scheduler/testplan/66c69c4008761ba27f76ed5d
https://elastictest.org/scheduler/testplan/66c69c2708761ba27f76ed5b
https://elastictest.org/scheduler/testplan/66cdafd1bd14ce56b2e820f7

Signed-off-by: xuliping <xuliping@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>