[dhcp_relay] Add dhcpv6 relay bind test when dhcpv6_server is not configured (#14164)
What is the motivation for this PR?
Add test for sonic-net/sonic-dhcp-relay#46
Add test to make sure dhcp6relay would not bind to vlan that doesn't configure dhcpv6 servers, and it wouldn't affect bind to other vlan

How did you do it?
Add test to make sure dhcp6relay would not bind to vlan that doesn't configure dhcpv6 servers, and it wouldn't affect bind to other vlan

Add a new vlan and assign ipv6 address to it. But not add dhcpv6 server for it
Check interface bind status
How did you verify/test it?
Run test

Signed-off-by: yenlu <yenlu@cisco.com>