Update the advanced reboot test to collect tcpdump on the server physical port
Get the tcpdump on the physical interface of the test server instead of the logical ptf interfaces. This is to fix the issue that sometimes there could be random packet drop on the ptf.
Keep the original approch to capture on the ptf interfaces for the kvm testbed which doesn't have the vmhost external port.